Creadas la carpeta pages y refactorizado el catálogo.
