+++ webapp/package.json
@@ -18,6 +18,7 @@
     "@types/react": "^17.0.37",
     "@types/react-dom": "^17.0.11",
     "express": "^4.17.2",
+    "mongoose": "^6.2.6",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-hook-form": "^7.27.1",
