+++ webapp/src/components/ProductosCatalogo/ProductoCatalogo/ProductoCatalogo.jsx
@@ -4,27 +4,27 @@ import useStyles from './styles'
 import { AddShoppingCart } from "@mui/icons-material";
 import { CardActionArea } from '@mui/material';
 
-const ProductoCatalogo = ({ product }) => {
+const ProductoCatalogo = ({ producto: producto }) => {
     const classes = useStyles();
 
     return (
         <Card className={classes.root}>
-            <CardActionArea href={'/' + product.id}>
-                <CardMedia className={classes.media} image={product.image} title={product.name} />
+            <CardActionArea href={'/' + producto.id}>
+                <CardMedia className={classes.media} image={producto.image} title={producto.name} />
                 <CardContent>
                     <div className={classes.cardContent}>
                         <Typography variant="h5" gutterButtom>
-                            {product.name}
+                            {producto.name}
                         </Typography>
 
                         <div className={classes.description}>
-                            <Typography variant="p" color="textSecondary">{product.description}</Typography>
+                            <Typography variant="p" color="textSecondary">{producto.description}</Typography>
                         </div>
                     </div>
                 </CardContent>
             </CardActionArea>
             <CardActions className={classes.cardActions}>
-                <Typography > {product.price} </Typography>
+                <Typography > {producto.price} </Typography>
                 <IconButton aria-label="Add to Cart">
                     <AddShoppingCart color='primary' fontSize='small' />
                 </IconButton>
